+++ libs/ui/src/lib/index.ts
@@ -1,5 +1,5 @@
 export * from './sidebar';
 export * from './components';
-export * from './models';
+export * from './modals';
 export * from './pages';
 export * from './Profile';
